New design and Responses component (#4)
* Prepared response objects for html layout

* Added styles for objects

* Put styles back in main.cs

* Initial refactoring for the css styles.
Remove semantic-ui so we have more control over the styles.
Remove Museo, we'll use Source Sans Pro to be consistent with gorgias-docs, and have less overhead.
Clean-up the navigation styles and fix the index links.

* Remove the unused gorgias icons.

* Refactor and clean-up the `Grid-*` classes as `content-block`.
Uses `table-cell` instead of flexbox, for IE support.
Removes the `Grid-inside` class and the need for a blank right column even when we had no content there.

* More improvements to the content styles.

* Basic mobile support.

* Remove right-column background on large screens.
The two color background is on the `.content` parent.

* Created new Response component

* Redesign the Properties tables.

* Refactor the request parameters table so we can re-use the properties component for it.

* Update the openapi.json file with the latest version from `/gorgias`.

* Improvements to the Responses component.
Fix issues with maps used to generate the reponse row items.

* Add an eslint-preloader, to lint as we go.
Fix linting errors everywhere.

* Improvements to the styles and markup for the Response object table.